asset: move common bootstrap ignition code into a helper type
Change the Bootstrap and BootstrapInPlace assets to compose a
Common type in order to isolate the in-place specific code.

Add the single-node ignition config target as a subtest of the
TestCreatedAssetsAreNotDirty test.
